final tweeks under way
