+++ __tests__/fc.Math.test.ts
@@ -663,8 +663,6 @@ describe('math', () => {
   })
 
 
-
-
   describe('ComplexNumerical', () => {
     let complexNumerical: ComplexNumerical;
