+++ src/@numerical.classes/index.ts
@@ -1,3 +1,4 @@
 export * from './bigint';
 export * from './number';
 export * from './testClass';
+export * from './complex'
